pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "main_project"
version = "0.0.1"
description = "An awesome project"
authors = [
  { name = "Group 150", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.66.1",
    "dvc-gdrive>=3.0.1",
    "dvc-ssh>=4.2.2",
    "fastapi==0.115.6",
    "hydra-core>=1.3.2",
    "kaggle>=1.6.0",
    "matplotlib>=3.8.0",
    "scikit-learn>=1.3.0",
    "torch==2.6.0",
    "kagglehub>=0.4.0",
    "pillow>=12.1.0",
    "torchvision>=0.21.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "wandb>=0.23.1",
    "torch-tb-profiler>=0.4.3",
    "dvc-gs>=3.0.2",
    "torch-tb-profiler>=0.4.3",
    "python-multipart>=0.0.6",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "httpx==0.27.0",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
source = ["src/main_project"]
omit = [
    "tests/*",
    "*/tmp*",
    "*/_remote_module_non_scriptable.py",
    "src/main_project/train.py",
    "src/main_project/api.py",
    "src/main_project/evaluate.py",
    "src/main_project/visualize.py",
    "src/main_project/__init__.py"]

[tool.dvc]
['remote "storage"']
url = "ssh://s252362@login.hpc.dtu.dk/zhome/5c/2/222250/mlops-data"
